Split SitemapPageViewModel to satisfy file_length SwiftLint rule
Extract enums (SitemapPageError, CommandLifecycleSummary, SitemapInteractionSummary,
RowInteractionState) into SitemapPageViewModel+SupportTypes.swift, computed view-state
properties into SitemapPageViewModel+InteractionState.swift, and error classification
logic into SitemapPageViewModel+PageHandlingError.swift. Drop unused
Published.Publisher.stream() extension. Main file is now 951 non-blank/comment lines.
