+++ OpenHABCore/Tests/OpenHABCoreTests/SessionTaskChallengeTests.swift
