+++ openHAB/UI/SwiftUI/ViewExtension.swift
